requirements.txt
attrs==22.2.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
exceptiongroup==1.1.0
geographiclib==2.0
geopy==2.3.0
idna==3.4
iniconfig==2.0.0
packaging==23.0
pluggy==1.0.0
pytest==7.2.0
python-dotenv==1.0.0
requests==2.28.1
soupsieve==2.3.2.post1
tomli==2.0.1
urllib3==1.26.13
